# spec/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Habbit Runner API
  version: 1.0.0
components:
  securitySchemes:
    accessTokenCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: HttpOnly access-token cookie
    csrfTokenHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF header required for cookie-authenticated mutations
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        errorCode:
          type: string
...
    PushSubscriptionEndpointRequest:
      type: object
      required:
      - endpoint
      properties:
        endpoint:
          type: string
          pattern: \S
    PushSubscriptionKeys:
      type: object
      required:
      - p256dh
      - auth
      properties:
        p256dh:
          type: string
          pattern: \S
        auth:
          type: string
          pattern: \S
    PushSubscriptionRequest:
      type: object
      required:
      - endpoint
      - keys
      properties:
        endpoint:
          type: string
          pattern: \S
        keys:
          $ref: "#/components/schemas/PushSubscriptionKeys"
    SubscriptionStatusResponse:
      type: object
      properties:
        success:
          type: boolean
...
    VapidPublicKeyResponse:
      type: object
      properties:
        publicKey:
          type: string
...
paths:
...
  /notifications/subscribe:
    post:
      summary: Subscribe to push notifications
      description: Stores the browser push subscription for the authenticated user.
      tags:
      - Notifications
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSubscriptionRequest"
        required: true
      responses:
        "201":
          description: Subscription stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionStatusResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Subscription endpoint already owned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - accessTokenCookie: []
  /notifications/unsubscribe:
    delete:
      summary: Unsubscribe from push notifications
      description: Removes the stored browser push subscription for the authenticated
        user.
      tags:
      - Notifications
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSubscriptionEndpointRequest"
        required: true
      responses:
        "204":
          description: Subscription removed
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authentication required or endpoint not owned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - accessTokenCookie: []
  /notifications/vapid-public-key:
    get:
      summary: Get VAPID public key
      description: Returns the public VAPID key used by web push clients.
      tags:
      - Notifications
      responses:
        "200":
          description: VAPID public key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VapidPublicKeyResponse"
        "503":
          description: VAPID key missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value